refactor: move template/ and metadata parsers into mtui/test_reports/
Carves the 'test-report templates + their metadata parsers' cluster out of
mtui's top level into a focused mtui/test_reports/ package, per Step 6 of
the reorg plan.

Moves:
* mtui/template/testreport.py    -> mtui/test_reports/testreport.py
* mtui/template/nulltestreport.py -> mtui/test_reports/null_report.py
* mtui/template/obstestreport.py  -> mtui/test_reports/obs_report.py
* mtui/template/pitestreport.py   -> mtui/test_reports/pi_report.py
* mtui/template/sltestreport.py   -> mtui/test_reports/sl_report.py
* mtui/template/products/         -> mtui/test_reports/products/
* mtui/template/__init__.py body  -> mtui/test_reports/svn_io.py
  (TemplateIOError, TestReportAlreadyLoadedError, TemplateFormatError,
  svn_commit_testreport, testreport_svn_checkout — extracted as a
  module sibling rather than left in test_reports/__init__.py so the
  boundary is explicit and import sites name the leaf).

Merges three sibling parser modules into one cohesive file
(metadata_parsers.py, ~210 lines, well under the discomfort line):
* mtui/parsemeta.py     (ReducedMetadataParser)
* mtui/parsemetajson.py (JSONParser)
* mtui/repoparse.py     (obsrepoparse / slrepoparse / gitrepoparse /
                         reporepoparse / _parse_product)

These three were always imported together by every concrete TestReport
subclass; merging them puts the parsing strategies next to the report
classes that consume them.

Test changes mirror the source moves: seven leaf test files are renamed
to match their new source modules, the three legacy parser tests are
merged into tests/test_metadata_parsers.py (also renamed for
source-name parity), and patch() targets are rewritten to follow the
moved modules. Logger names ('mtui.template', 'mtui.template.testreport',
'mtui.template.sltestreport') are intentionally kept unchanged — logger
identifiers are stable external contracts independent of the import
path, per the precedent set in commit 1be141b. Tests asserting log
output via caplog continue to reference the old logger strings.
