+++ mtui/commands/approve.py
@@ -16,7 +16,7 @@ from ..cli.term import prompt_user
 from ..connector import OSC, Gitea
 from ..support.exceptions import GiteaError, InvalidGiteaHashError
 from ..support.misc import requires_update
-from ..template import TemplateFormatError, svn_commit_testreport
+from ..test_reports.svn_io import TemplateFormatError, svn_commit_testreport
 from .apicall import BaseApiCall
 
 logger = getLogger("mtui.command.approve")
