+++ mtui/test_reports/obs_report.py
@@ -1,28 +1,26 @@
-"""A `TestReport` implementation for PI test reports."""
+"""A `TestReport` implementation for OBS test reports."""
 
 from typing import final
 
-from ..parsemeta import ReducedMetadataParser
-from ..parsemetajson import JSONParser
-from ..repoparse import reporepoparse
 from ..target import Target
 from ..target.hostgroup import HostsGroup
 from ..types import Product, RequestReviewID
+from .metadata_parsers import JSONParser, ReducedMetadataParser, obsrepoparse
 from .testreport import TestReport
 
 
 @final
-class PITestReport(TestReport):
-    """A `TestReport` implementation for PI test reports."""
+class OBSTestReport(TestReport):
+    """A `TestReport` implementation for OBS test reports."""
 
     def __init__(self, *a, **kw) -> None:
-        """Initializes the `PITestReport` object."""
+        """Initializes the `OBSTestReport` object."""
         super().__init__(*a, **kw)
 
         self.rrid: RequestReviewID
         self.rating = ""
         self.realid = ""
-        self.repositories: frozenset = frozenset()
+
         self._attrs += ["rrid", "rating", "realid"]
 
     @property
@@ -40,19 +38,16 @@ class PITestReport(TestReport):
 
     def _update_repos_parser(self) -> dict[Product, str]:
         """Returns a dictionary of update repositories."""
-        return reporepoparse(self.repositories, self.products)
+        return obsrepoparse(self.repository, self.report_wd())
 
     def _show_yourself_data(self) -> list[tuple[str, str]]:
         """Returns a list of data to be displayed by `list_metadata`."""
-        return (
-            [
-                ("ReviewRequestID", str(self.rrid)),
-                ("Rating", self.rating),
-                ("Real ID", self.realid),
-            ]
-            + [("Repo", x) for x in self.repositories]
-            + super()._show_yourself_data()
-        )
+        return [
+            ("ReviewRequestID", str(self.rrid)),
+            ("Rating", self.rating),
+            ("Real ID", self.realid),
+            *super()._show_yourself_data(),
+        ]
 
     def set_repo(self, target: Target, operation: str) -> None:
         """Adds or removes a repository on a target host.
@@ -63,7 +58,7 @@ class PITestReport(TestReport):
 
         """
         if operation == "add":
-            target.repo_manager.run_zypper("-n ar -cfGkn", self.update_repos, self.rrid)
+            target.repo_manager.run_zypper("-n ar -ckn", self.update_repos, self.rrid)
         elif operation == "remove":
             target.repo_manager.run_zypper("-n rr", self.update_repos, self.rrid)
         else:
